// next-movie/types/global.ts
export type GenreType = {
    id: string;
    name: string;
}

export type MovieType = {
    id: string;
    title: string;
    overview: string;
    poster_path: string;
    backdrop_path: string;
    release_date: string;
}

export type PersonType = {
    id: string;
    name: string;
    character: string;
    profile_path: string;
}